pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bot-gpt-backend"
version = "0.1.0"
description = "Backend for BOT GPT conversational AI platform with RAG capabilities"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { text = "MIT" }
keywords = ["fastapi", "llm", "rag", "faiss", "conversational-ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Web Framework
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    # Database
    "sqlalchemy==2.0.23",
    # Data Validation
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    # Vector Search & Embeddings
    "faiss-cpu==1.7.4",
    "sentence-transformers==2.2.2",
    # HTTP & Utilities
    "requests==2.31.0",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "httpx==0.25.1",
    # Code Quality
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.1",
]

gpu = [
    # GPU-accelerated FAISS (optional)
    "faiss-gpu==1.7.4",
]

[project.urls]
Homepage = "https://github.com/yourusername/bot-gpt-backend"
Documentation = "https://github.com/yourusername/bot-gpt-backend#readme"
Repository = "https://github.com/yourusername/bot-gpt-backend.git"
Issues = "https://github.com/yourusername/bot-gpt-backend/issues"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
